Make CLI entrypoint a thin dispatcher
