+++ packages/cli/src/commands/exec.ts
@@ -1,6 +1,9 @@
-import { attach, formatMachinenError, isMachinenError, type VmHandle } from "@machinen/runtime";
+import { attach, type VmHandle } from "@machinen/runtime";
 
-import { extractTarget, type Target } from "../parse-target.ts";
+import { die, handleError } from "../errors.ts";
+import type { Target } from "../parse-target.ts";
+import { runPtyExec } from "./pty.ts";
+import { parseTargetFlags } from "./target.ts";
 
 export async function cmdExec(args: string[]): Promise<number> {
   const parsed = parseExecArgs(args);
@@ -64,69 +67,4 @@ async function runRawExec(vm: VmHandle, cmd: string): Promise<number> {
     onStderr: (chunk) => process.stderr.write(chunk),
   });
   return res.exitCode;
-}
-
-async function runPtyExec(vm: VmHandle, cmd: string): Promise<number> {
-  const tty = enterPtyRawMode();
-  const handle = vm.execPty(cmd, {
-    cols: tty.cols,
-    rows: tty.rows,
-    stdin: process.stdin,
-    stdout: process.stdout,
-  });
-  const onResize = () =>
-    handle.resize(process.stdout.columns ?? tty.cols, process.stdout.rows ?? tty.rows);
-  process.stdout.on("resize", onResize);
-  try {
-    const { exitCode } = await handle.result;
-    return exitCode;
-  } finally {
-    process.stdout.removeListener("resize", onResize);
-    tty.restore();
-  }
-}
-
-function enterPtyRawMode(): { cols: number; rows: number; restore: () => void } {
-  const wasRaw = process.stdin.isRaw === true;
-  process.stdin.setRawMode(true);
-  process.stdin.resume();
-  return {
-    cols: process.stdout.columns ?? 80,
-    rows: process.stdout.rows ?? 24,
-    restore: () => restorePtyRawMode(wasRaw),
-  };
-}
-
-function restorePtyRawMode(wasRaw: boolean): void {
-  if (wasRaw) {
-    return;
-  }
-  try {
-    process.stdin.setRawMode(false);
-  } catch {}
-}
-
-function parseTargetFlags(args: string[], cmd: string): Target {
-  try {
-    const { target, rest } = extractTarget(args, cmd);
-    if (rest.length > 0) {
-      die(`unknown argument: ${rest[0]}`);
-    }
-    return target;
-  } catch (err) {
-    handleError(err);
-  }
-}
-
-function handleError(err: unknown): never {
-  if (isMachinenError(err)) {
-    process.stderr.write(`machinen: ${formatMachinenError(err)}\n`);
-    process.exit(1);
-  }
-  throw err;
-}
-
-function die(msg: string): never {
-  process.stderr.write(`machinen: ${msg}\n`);
-  process.exit(1);
 }
